refactor: centralize move generation
